+++ cartridges/death_star/.project
@@ -1,0 +1,17 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<projectDescription>
+    <name>death_star</name>
+    <comment></comment>
+    <projects>
+    </projects>
+    <buildSpec>
+        <buildCommand>
+            <name>com.demandware.studio.core.beehiveElementBuilder</name>
+            <arguments>
+            </arguments>
+        </buildCommand>
+    </buildSpec>
+    <natures>
+        <nature>com.demandware.studio.core.beehiveNature</nature>
+    </natures>
+</projectDescription>
