+++ cartridges/death_star/.tern-project
@@ -1,0 +1,11 @@
+{
+    'ecmaVersion': 5,
+    'plugins': {
+        'guess-types': {
+        },
+        'outline': {
+        },
+        'demandware': {
+        }
+    }
+}'
