+++ cartridges/death_star/cartridge/death_star.properties
@@ -1,0 +1,4 @@
+## cartridge.properties for cartridge death_star
+#Mon Nov 20 14:45:1 CEST 2021
+demandware.cartridges.death_star.multipleLanguageStorefront=true
+demandware.cartridges.death_star.id=death_star
