+++ cartridges/death_star/cartridge/templates/death.isml
@@ -1,0 +1,7 @@
+AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+
+<div class="container">
+    
+    <h2 class='text-center'>TEST </h2>
+    
+</div>
